refactor: props 타입 파일에 포함 (#103)
