+++ src/pages/match/components/mate-carousel.tsx
@@ -1,9 +1,16 @@
 import { cn } from '@libs/cn';
 import CarouselIndicator from '@pages/match/groups/components/carousel_indicator';
+import type { DetailedCardProps } from '@components/card/match-card/types/card';
 import SlideItem from '@pages/match/groups/components/slide-item';
-import type { MateCarouselProps } from '@pages/match/groups/types/carousel-indicator';
 import { useSlide } from '@pages/match/hooks/useSlide';
 import { getSlideTransformStyle } from '@pages/match/styles/get-slide-transformstyle';
+
+interface MateCarouselProps {
+  mates: (DetailedCardProps & { id: number })[];
+  currentIndex: number;
+  onDotClick: (index: number) => void;
+  isGroupMatching: boolean;
+}
 
 const MateCarousel = ({ mates, currentIndex, onDotClick, isGroupMatching }: MateCarouselProps) => {
   const { handleTouchStart, handleTouchEnd, handleMouseDown } = useSlide({
