+++ src/pages/match/components/mate-header.tsx
@@ -1,5 +1,9 @@
 import { MATCHING_HEADER_MESSAGE } from '@pages/match/constants/matching';
-import type { MateHeaderProps } from '@pages/match/groups/types/mate-type';
+
+interface MateHeaderProps {
+  nickname: string;
+  isGroupMatching?: boolean;
+}
 
 const MateHeader = ({ nickname = '사용자', isGroupMatching }: MateHeaderProps) => (
   <section className="gap-[0.8rem] text-center">
