+++ src/pages/match/components/mate.tsx
@@ -1,9 +1,14 @@
 import MateCarousel from '@pages/match/components/mate-carousel';
 import MateFooter from '@pages/match/components/mate-footer';
 import MateHeader from '@pages/match/components/mate-header';
-import type { MateProps } from '@pages/match/groups/types/mate-type';
 import useMate from '@pages/match/hooks/useMate';
 import { useState } from 'react';
+
+interface MateProps {
+  matchId: number;
+  onRequestClick: () => void;
+  isGroupMatching?: boolean;
+}
 
 const Mate = ({ matchId, onRequestClick, isGroupMatching = true }: MateProps) => {
   const { mates } = useMate(matchId);
