+++ src/pages/result/components/sent-view.tsx
@@ -1,10 +1,14 @@
 import Button from '@components/button/button/button';
 import { LOTTIE_PATH } from '@constants/lotties';
 import { MATCHING_COMPLETE_MESSAGE } from '@pages/match/constants/matching';
-import type { SentViewProps } from '@pages/match/groups/types/mate-type';
 import { ROUTES } from '@routes/routes-config';
 import { Lottie } from '@toss/lottie';
 import { useNavigate } from 'react-router-dom';
+
+interface SentViewProps {
+  isGroupMatching?: boolean;
+  userNickname: string;
+}
 
 const SentView = ({ isGroupMatching = true, userNickname }: SentViewProps) => {
   const navigate = useNavigate();
